patch 8.0.0746: when :term fails the job is not properly cleaned up
Problem:    When :term fails the job is not properly cleaned up.
Solution:   Free the terminal. Handle a job that failed to start. (closes
            #1858)

+++ src/os_unix.c
@@ -5475,7 +5475,9 @@ mch_stop_job(job_T *job, char_u *how)
 	job_pid = -job_pid;
 #endif
 
-    kill(job_pid, sig);
+    /* Never kill ourselves! */
+    if (job_pid != 0)
+	kill(job_pid, sig);
 
     return OK;
 }
